+++ cuba_weather_owm/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.0.6'
+__version__ = '0.0.7'
 
 from cuba_weather_municipality import CubaWeatherMunicipality
 from .models import WeatherModel